Add files via upload
In this project, I used Python to build a model that can predict the price of a house based on different factors. I used Python's libraries like NumPy, Pandas, and Matplot. These libraries helped me to process, clean, balance the skewness of features, and perform feature engineering. Then, I trained a machine learning algorithm using the cleaned data to create the prediction model and tried several other parameters except the default one to find the best parameters for the model.